+++ src/types/types.ts
@@ -60,3 +60,20 @@ export interface ChatInfo {
     content: string;
   }
 }
+
+export interface Message {
+  chat_id: number;
+  time: string;
+  type: string;
+  user_id: number;
+  content: string;
+  file?: {
+    id: number;
+    user_id: number;
+    path: string;
+    filename: string;
+    content_type: string;
+    content_size: number;
+    upload_date: string;
+  }
+}
